ANP-23 Refactored TopAppBar, SpinningLogo, TopAppBarLogoTitle into their own components for greater reusability
