+++ android/app/src/main/java/hr/foi/air/wattsup/pages/ChargerPage.kt
@@ -7,14 +7,12 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -22,26 +20,20 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import hr.foi.air.wattsup.R
+import hr.foi.air.wattsup.ui.component.TopAppBar
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun ChargerPage(onArrowBackClick: () -> Unit) {
     Scaffold(
         topBar = {
-            CenterAlignedTopAppBar(
-                title = {
-                    Text(stringResource(R.string.charger_mode))
-                },
-                colors = TopAppBarDefaults.mediumTopAppBarColors(
-                    containerColor = MaterialTheme.colorScheme.primaryContainer,
-                    titleContentColor = MaterialTheme.colorScheme.onPrimary,
-                ),
+            TopAppBar(
+                title = { Text(stringResource(R.string.charger_mode)) },
                 navigationIcon = {
                     IconButton(onClick = { onArrowBackClick() }) {
                         Icon(Icons.Filled.ArrowBack, null, tint = Color.White)
                     }
                 },
-
             )
         },
     ) {
@@ -64,5 +56,5 @@ fun ChargerPage(onArrowBackClick: () -> Unit) {
 @Preview
 @Composable
 fun ChargerPagePreview() {
-    ChargerPage({})
+    ChargerPage {}
 }
