+++ android/app/src/main/java/hr/foi/air/wattsup/ui/component/SpinningLogo.kt
@@ -1,0 +1,38 @@
+package hr.foi.air.wattsup.ui.component
+
+import android.os.Build
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import coil.ImageLoader
+import coil.compose.rememberAsyncImagePainter
+import coil.decode.GifDecoder
+import coil.decode.ImageDecoderDecoder
+import coil.request.ImageRequest
+import hr.foi.air.wattsup.R
+
+@Composable
+fun SpinningLogo() {
+    val context = LocalContext.current
+    val imageLoader = ImageLoader.Builder(context)
+        .components {
+            if (Build.VERSION.SDK_INT >= 28) {
+                add(ImageDecoderDecoder.Factory())
+            } else {
+                add(GifDecoder.Factory())
+            }
+        }
+        .build()
+    Image(
+        painter = rememberAsyncImagePainter(
+            ImageRequest.Builder(context).data(data = R.drawable.icon_white).apply(block = {
+                size(150, 150)
+            }).build(),
+            imageLoader = imageLoader,
+        ),
+        contentDescription = null,
+        modifier = Modifier.fillMaxWidth(),
+    )
+}
