+++ android/app/src/main/java/hr/foi/air/wattsup/ui/component/TopAppBar.kt
@@ -1,0 +1,24 @@
+package hr.foi.air.wattsup.ui.component
+
+import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun TopAppBar(title: @Composable () -> Unit, navigationIcon: @Composable () -> Unit) {
+    CenterAlignedTopAppBar(
+        title = {
+            title()
+        },
+        colors = TopAppBarDefaults.mediumTopAppBarColors(
+            containerColor = MaterialTheme.colorScheme.primaryContainer,
+            titleContentColor = MaterialTheme.colorScheme.onPrimary,
+        ),
+        navigationIcon = {
+            navigationIcon()
+        },
+    )
+}
